requirements.txt
torch
numpy
transformers
tqdm
pillow